+++ apps/server/app.ts
@@ -9,10 +9,7 @@ import { createServerRouter } from "@/library/router";
 import type { ApiRequestWithUser } from "@/library/types/api.types";
 import emailClient from "@/library/ses.client";
 
-import {
-	AuthController,
-	UsersController
-} from "./routes";
+import * as controllers from "./controllers";
 
 import { morganMiddleware, createIdentificationMiddleware } from "@/library/middlewares";
 
@@ -43,14 +40,12 @@ container.register<EmailService>(
 
 const authServiceInstance = container.resolve(AuthService);
 
-const controllers = [
-	AuthController,
-	UsersController
-]
-
 const middlewares = [
 	createIdentificationMiddleware(authServiceInstance),
 	morganMiddleware
 ]
 
-export default createServerRouter<ApiRequestWithUser>({ controllers, middlewares });
+export default createServerRouter<ApiRequestWithUser>({
+	controllers: Object.values(controllers),
+	middlewares
+});
